fix(tests): Update unit and functional tests to handle throttled password reset email requests.
* Merge all reset password related functional tests into one file that tests the full flow.

fixes #842
