+++ tests/functional.js
@@ -11,9 +11,6 @@ define([
   './functional/pp',
   './functional/confirm',
   './functional/reset_password',
-  './functional/confirm_reset_password',
-  './functional/complete_reset_password',
-  './functional/reset_password_complete',
   './functional/robots_txt',
   './functional/settings',
   './functional/force_auth',
